pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Flask app to analyze pages"
authors = ["Tve1n hexlet studen"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
